+++ gdsfactory/pack.py
@@ -106,10 +106,13 @@ def pack(
     density: float = 1.1,
     precision: float = 1e-2,
     text: Optional[ComponentFactory] = None,
-    prefix: str = "",
-    offset: Float2 = (0, 0),
-    anchor: Anchor = "cc",
+    text_prefix: str = "",
+    text_offset: Float2 = (0, 0),
+    text_anchor: Anchor = "cc",
     name_prefix: Optional[str] = None,
+    rotation: int = 0,
+    h_mirror: bool = False,
+    v_mirror: bool = False,
 ) -> List[Component]:
     """Pack a list of components into as few Components as possible.
 
@@ -123,10 +126,15 @@ def pack(
         sort_by_area: Pre-sorts the shapes by area
         density: Values closer to 1 pack tighter but require more computation
         precision: Desired precision for rounding vertex coordinates.
-        text: function for creating the text
-        prefix: for the text
-        offset: relative to component center
-        anchor: size info (ce cw nc ne nw sc se sw center cc)
+        text: Optional function to add text labels.
+        text_prefix: for labels. For example. 'A' will produce 'A1', 'A2', ...
+        text_anchor: size info (ce cw nc ne nw sc se sw center cc). Defaults to center.
+        text_offset: relative to component size info anchor. Defaults to center.
+        name_prefix: for each packed component (avoids the Unnamed cells warning).
+            Note that the suffix contains a uuid so the name will not be deterministic
+        rotation: for each component in degrees
+        h_mirror: horizontal mirror using y axis (x, 1) (1, 0). This is the most common mirror.
+        v_mirror: vertical mirror using x axis (1, y) (0, y)
     """
     i = 0
 
@@ -180,14 +188,15 @@ def pack(
             xcenter = x + w / 2 + spacing / 2
             ycenter = y + h / 2 + spacing / 2
             component = component_list[n]
-            d = packed.add_ref(component)
+            d = component.ref(rotation=rotation, h_mirror=h_mirror, v_mirror=v_mirror)
+            packed.add(d)
             if hasattr(component, "settings"):
                 packed.info["components"][component.name] = component.settings
             d.center = (xcenter * precision, ycenter * precision)
 
             if text:
-                label = packed << text(f"{prefix}{i}")
-                label.move((np.array(offset) + getattr(d.size_info, anchor)))
+                label = packed << text(f"{text_prefix}{i}")
+                label.move((np.array(text_offset) + getattr(d.size_info, text_anchor)))
                 i += 1
 
         components_packed_list.append(packed)
@@ -258,12 +267,15 @@ if __name__ == "__main__":
     # c.write_gds_with_metadata("mask.gds")
 
     p = pack(
-        [gf.components.rectangle(size=(i, i), port_type=None) for i in range(1, 10)],
+        [gf.components.triangle(x=i) for i in range(1, 10)],
         spacing=20.0,
         max_size=(100, 100),
-        text=gf.c.text,
-        prefix="R",
-        # name_prefix="demo",
+        text=gf.partial(gf.c.text, justify="center"),
+        text_prefix="R",
+        name_prefix="demo",
+        text_anchor="nc",
+        text_offset=(-10, 0),
+        v_mirror=True,
     )
     c = p[0]
     print(c.name)